# fix: Optimizes GetString to eliminate allocations